Various improvements, UI WIP
